Dev doctor reservation dy

See merge request s08-webmobile1-sub2/S08P12A304!78

+++ Backend/marathon/src/main/java/com/ssafy/marathon/dto/response/treatment/TreatmentResDto.java
@@ -15,7 +15,7 @@ import lombok.ToString;
 @NoArgsConstructor
 @AllArgsConstructor
 public class TreatmentResDto {
-    private int firstDateInfo;
+    private Long firstDateInfo;
 
     private List<DayOfTreatmentResDto> list;
